requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

onnxruntime==1.23.2
numpy==1.24.3
Pillow==10.1.0

torch==2.1.1
torchvision==0.16.1
scikit-learn==1.3.2

matplotlib==3.8.2
seaborn==0.13.0

pytest==7.4.3
httpx==0.25.2
